requirements.txt
django==1.8.3
psycopg2==2.6.1
djangorestframework==3.1.3
